Github Actions tests of playbook (#1)
This introduces a testsuite for the playbook run itself on Ubuntu 22.04 and 20.04 for three python versions.

re: .github/workflows/ansible-playbook.yml
The ansible-galaxy install randomly fails in actions. Using nick-fields/retry@v2 will retry the install thrice before failing for just this specific step.

re: tasks/mysql.yml
As the Github Actions Ubuntu images include a mysql server, the user/password for this has been added as an "omit" parameter to the mysql tasks.

re: tests/test-gha.yml
Since its a very minimal setup of snipe, the default vhost of the nginx install is kept at localhost, which also means that we need to remove the default.conf that is shipped with nginx to actually be able to reach our snipe-it install via localhost.